Check for the file/alias before attempting to create/clone

+++ bootstrap.sh
@@ -4,20 +4,32 @@ if [ ! -d ~/.vim/bundle ]; then
 	mkdir ~/.vim/bundle
 	cd ~/.vim/bundle
 
-	git clone https://github.com/tomtom/tlib_vim.git
-	git clone https://github.com/MarcWeber/vim-addon-mw-utils.git
-	git clone https://github.com/SirVer/ultisnips.git
-	git clone https://github.com/honza/vim-snippets.git
-	git clone https://github.com/rodjek/vim-puppet.git
-	git clone https://github.com/scrooloose/nerdtree.git
-	git clone https://github.com/skrzyp/vim-mads
-	git clone https://github.com/tpope/vim-surround.git
-	git clone https://github.com/reedes/vim-pencil
-	git clone https://github.com/djoshea/vim-autoread.git
-	git clone https://github.com/junegunn/goyo.vim.git
-	git clone https://github.com/plasticboy/vim-markdown.git
+    echo "Cloning tlib..."
+	git clone --quiet https://github.com/tomtom/tlib_vim.git
+    echo "Cloning vim-addon-mw-utils..."
+	git clone --quiet https://github.com/MarcWeber/vim-addon-mw-utils.git
+    echo "Cloning Ultisnippets..."
+	git clone --quiet https://github.com/SirVer/ultisnips.git
+    echo "Cloning vim-snippets..."
+	git clone --quiet https://github.com/honza/vim-snippets.git
+    echo "Cloning vim-puppet..."
+	git clone --quiet https://github.com/rodjek/vim-puppet.git
+    echo "Cloning nerdtree..."
+	git clone --quiet https://github.com/scrooloose/nerdtree.git
+    echo "Cloning vim-mads..."
+	git clone --quiet https://github.com/skrzyp/vim-mads
+    echo "Cloning vim-surround..."
+	git clone --quiet https://github.com/tpope/vim-surround.git
+    echo "Cloning vim-pencil..."
+	git clone --quiet https://github.com/reedes/vim-pencil
+    echo "Cloning vim-autoread..."
+	git clone --quiet https://github.com/djoshea/vim-autoread.git
+    echo "Cloning Goyo..."
+	git clone --quiet https://github.com/junegunn/goyo.vim.git
+    echo "Cloning vim-markdown..."
+	git clone --quiet https://github.com/plasticboy/vim-markdown.git
 else
 	echo "~/.vim/bundle folder already exists..."
 fi
 
-ln -s ~/.vim/.vimrc ~/.vimrc
+[ ! -L ~/.vimrc ] && ln -s ~/.vim/.vimrc ~/.vimrc || echo "Symlink already exists..."
